build: Add build mode check for common libraries
Add a macro OPAL_MCA_CHECK_DEPENDENCY() to check if a component's
build mode (static / dso) conflicts with a dependent component's
build mode (generally, a common library).  The issue of concern
is if the component is built as a static component, but the
common component is build as a dso.  Since all the common components
currently depend on libopen-pal (and sometimes libmpi), libopen-pal
and/or libmpi cannot depend on a common library without creating a
circular dependency chain, and we do not like those.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>