build: Add m4 macro to assert argument is literal
Add a new m4 macro, OAC_ASSERT_LITERAL, that will fail autoconf if
the first argument is not a string literal (in the Autoconf sense
of the word).  Note that this macro will not emit any shell code;
it will either cause a fatal m4 error (with useful stack trace
to where the error occurred) or evaluate to an empty string.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit edcf4cac6739ef214d75fc40bcd604c5089be342)