build: Refactor OPAL_VAR_SCOPE system
Refactor the OPAL_VAR_SCOPE_PUSH and OPAL_VAR_SCOPE_POP macros to
be wrappers around shell functions to reduce code duplication in
the output configure.  This patch reduces the size of the output
configrue by approximately 30% on MacOS.

Move the scope stack from shell to m4, as all the inputs are
currently string literals, so we can have the same functionality
with less work in the resulting configure script.

With the scope stack in m4, also perform variable name conflicts
during autogen time.  This allows for us to detect more conflicts,
as there were times that the parent macro pushed a variable
name in the scope but didn't initialize it before calling a
child macro with a conflict, which was a bit of a ticking timebomb.
Note that we do still search the current shell environment during
OPAL_VAR_SCOPE_PUSH, as macros may use a variable without pushing
it into a scope.

FInally, be more aggressive about cleaning up the environment after
ourselves.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit a8c829c5899ff768878c9ab4e310bf23bb6c08e7)