PMIx_Group_construct - handle error returns
The error return code from PMIx_Group_construct was not being handled at all.
This PR fixes this issue.

In cases where the process is running without a PMIx server, for instance in a singleton
program which has not invoked MPI_Comm_spawn, and it uses MPI_Comm_create_from_group
the following error hep message is emitted:

--------------------------------------------------------------------------

Your application has invoked an MPI function that is not supported in
this environment.

  MPI function: MPI_Comm_from_group/MPI_Intercomm_from_groups
  Reason:       PMIx server unreachable

--------------------------------------------------------------------------

If a PMIx implementation is being used that does not support PMIx_Group_construct,
the following help message is emitted if the code invokes MPI_Comm_create_from_group
or MPI_Intercomm_create_from_groups:

--------------------------------------------------------------------------

Your application has invoked an MPI function that is not supported in
this environment.

  MPI function: MPI_Comm_from_group/MPI_Intercomm_from_groups
  Reason:       PMIx server does not support PMIx Group operations

--------------------------------------------------------------------------

Related to #10736

Signed-off-by: Howard Pritchard <howardp@lanl.gov>