pr-checks: use reusable actions for commit checker and labeler
An OMPI maintainer wanted the commit checker and labeler in two of their other
repos. Instead of copying code, it's far easier to move the actions to separate
repos and use them as reusable actions. They now live at:

https://github.com/open-mpi/pr-git-commit-checker
https://github.com/open-mpi/pr-labeler

This means any development on these (and new, future) actions can happen in
their dedicated repos. Additionally, the actions can easily be used by anyone
else on GitHub. For convenience's sake, the action calls here could also be
setup to track a branch (such as v1) instead of the release (v1.0.0, as it is
now). Therefore, if changes are made to the action upstream, it need not be
updated here. However, right now, if OMPI wants to make use of upstream changes,
the YAML needs to be changed here as well.

The functionality of the actions remains the same, but they are now called
slightly differently, using inputs to pass config and the GITHUB_TOKEN secret
into the actions. Instead of using JSON to make cherry picks required on other
branches, you pass it in as an input.

Signed-off-by: Joe Downs <joe@dwns.dev>