Avoid name collisions with intrinsics functions
Many compilers (tested with gcc and clang) use the memcpy and memmove
keywords as intrinsics functions. They also lack a proper syntactic
matching, and this prevents the use of any intrincs names as members
of structures. Use a different name for the 2 members of the accelerator
framework that handles memory copies and moves.

Fixes #10869.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>