opal/event: use epoll by default on Linux
Under normal circumstances epoll and poll produce similar performance on Linux.
When busy polling is enabled they do not. Testing with a TCP-based system shows
a significan performance degredation when using poll with busy waiting enabled.
This performance regression is not seen when using epoll. This PR adjusts the
default value of opal_event_include to epoll on Linux only to fix the
regression.

Fixes #10929

Signed-off-by: Nathan Hjelm <hjelmn@google.com>