add accelerator and parts framework to mca lists
This patch adds part framework to the list of ompi frameworks opened when the core of open mpi initialized and
closed when the ref cnt on the ompi core drops back to zero.

Likewise, the accelerator framework is added to the opal frameworks list.

Note the accelerator framework cleanup function does need to be added to opal's cleanup list.

Add some verbiage that will hopefully show where to add new frameworks in the opal/ompi initialization procedures.

Add a macro in opal_init.c that generates complilation errors if a call to mca_base_framework_open is made with
within opal_init.

Also add a cleanup function for the smcuda btl to avoid a warning message at mpi finalize.

Related to #10938

Signed-off-by: Howard Pritchard <howardp@lanl.gov>