Fix type of returned error code for get_dynamic_win_info
The get_dynamic_win_info function is in the ompi namespace and
should therefore return OMPI_* error codes. The 'int ret' variable should
only be used to store the final return value of the function, here.
To cover errors in the unlock call in the cleanup code use 'ret_unlock' -
but let previous errors take precedence over it.

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>