singleton: reduce chattiness under slurm
The changes to address issue #10286 had the side effect of adding
some unexpected output in the case an application is being launched
in singleton mode and happens to be within a slurm allocation.

This patch is more selecive in the environment variables checked to
determine whether or not a process was launched under srun vs just
being a process that happens to be running within a slurm allocation.

Related to #11034

Signed-off-by: Howard Pritchard <howardp@lanl.gov>