mca/base CID 1516779: fix possible NULL usage
In practice, this probably would never happen because we'll always
have a path with a path separator in it.  But we might as well be
defensive and ensure to handle the case where we strtok() returns NULL
on even the first call.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>
(cherry picked from commit 02add5fef508e7a09caccc59fe17bb9aa6a9a2cd)