opal/util/ethtool.c: fix interface name len
Per a comment in opal/util/if.h, we hard-code OPAL_IF_NAMESIZE to 32 --
regardless of what the underlying IF_NAMESIZE is.  On some Linux
systems, IF_NAMESIZE if 16, leading to a potential memory overrun when
we copy the interface name.  In practice, this probably wouldn't
happen because we use opal_string_copy(), which will only copy up to
the terminating \0, and Linux network interface names will always have
a terminating \0.

But we might as well be defensive, and use the underlying IF_NAMESIZE
(this file already includes <net/if.h>, where that should be defined).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>