mtl: Fix datatype offsetting
The convertor packing code will automatically add
datatype offsets. However, the MTL avoids entering the
convertor in some cases and thus was required to add
an offset somewhere. The original location -
mca_pml_cm_send() - was causing a problem with accelerator
buffers as it would add the offset before sending it to
the opal_convertor_pack function, doubling the offset.

The correct place to add the datatype offset was modified
to be in the ompi_mtl_datatype_pack() function which
appropriately mirrors the opal_convertor_pack() offset
behavior.

Signed-off-by: William Zhang <wilzhang@amazon.com>