OFI/MTL: bypass HMEM mem reg for CXI provider
The code in the OFI MTL to register acclerator memory doesn't work well with providers like CXI which require the libfabric
consumer to specify the requested_key field in the fi_mr_attr arg to fi_mr_regattr.  As it turns out there's a simpe workaround
though with this provider as it actually doesn't need accelerator memory to be registered for FI_MSG/FI_TAGGED type transfers.

This patch restores support for transfers to/from accelerator memory using the OFI MTL on SS11 (CXI provider).

Signed-off-by: Howard Pritchard <howardp@lanl.gov>