test/datatype/partial.c: fix compiler warnings
There are a bunch of other warnings in tests, but I was recently
working with a user in this specific test, and the warnings annoyed
me.  So I fixed them in this test (but not other tests).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>
(cherry picked from commit 98abc08a6e2a0a4181a823c1b81a238964710240)