io/romio341: fix support for GCC 4.8 compilers
ROMIO 3.4.1 requires <stdatomic.h>, but this is not checked by
the embedded configure, causing a failure at compilation time
if the compiler (such as GCC 4.8 provided by RHEL7 like distros).

Explicitly check for <stdatomic.h> and disable the io/romio341
component and the third party ROMIO if not available.

Refs. open-mpi/ompi#11088

Signed-off-by: Gilles Gouaillardet <gilles@rist.or.jp>