common/ompio: add pipelined file_iwrite and iread
Pipelined iread/iwrite operations require the notion of subrequests, i.e. a user level request
can contain multiple internal subrequests that all have to complete before the user level operation
is considered finished. This requires adjustments
to the internal ompio progress engine and data structures.

Note: this is purely just a pipelined algorithm, no overlap between different iterations.

Extract the file view into a separate datastructure. This is required in
the next step since we need to cache file view and position of the file pointer on the request.

Replicate the file view and the file pointer position on the request.
This is required to correctly increment where to read/write data for every
subrequest, and handle the potential situation that the code changes the file
after posting an iread/iwrite but before the operation finishes.

Furthermore, when initiating an iread/iwrite we need to immidiatly move the position
of the handle to the end of the operation, such that subsequent read/write operations
use the correct position to start out with, and don't accidentaly interfere
with the already ongoing non-blocking operation.
Signed-off-by: Edgar Gabriel <edgar.gabriel@amd.com>

Signed-off-by: Edgar Gabriel <edgar.gabriel@amd.com>