remove redundant PMIx fence calls at initialization
related to issue #11166

Turns out making these changes revealed some issues with the way
MPI_Session_init is implemented.  Unless async modex and disabling
of collective info is turned on, the path through MPI initialization
for Sessions isn't really local.  The code in ompi_mpi_instance_init_common
will eventually need to be refactored to move synchronizing elements
up into ompi_mpi_init for the world process model and into some delayed-till-
comm creation routine(s) when using the sessions process model.

also, remove a second wait on a debugger.  We should only do this
once.  At some point a solution for attaching a debugger to an application
which only uses the sessions process model will need to implemented.
Given the semantics of sessions, its not clear when is the right point to
synchronize processes with a debugger.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>