Fix singleton operations
You cannot daemonize the "prte" executable when spawning it
to support a singleton as that will cause things to hang. Also
fix IO forwarding thru the singleton for the spawned child
procs by correcting a mistake that caused the IOF request
attributes to be overlooked when constructing the job info for
the PMIx_Spawn call.

Includes an update to the PMIx and PRRTE submodule pointers to
pickup a couple of relevant corrections there. See:

https://github.com/openpmix/prrte/pull/1621
https://github.com/openpmix/openpmix/pull/2881

This brings the submodule pointers to the HEAD of their respective
release branches, which are basically at an rc1 level (but not
tagged yet).

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 8ca4d7c2d9c51860cfe0b3c0c2edab61be450273)