coll/han: reorder free calls and avoid read-after-free in debug builds
Inside the main loop, whenever we read a new string we free it first.
The first iteration will be free(NULL), which is legal.
At the end, we free all strings in all paths.
This removes a potential read-after-free in a debug build and removes
some calls to free from the error paths.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>