accelerator/rocm: add buffer_id
add functionality to retrieve a buffer id. This is required for the
rcache module to identify whether a buffer pointer is still the
one that has originally been registered.

In addition, it seems that I forgot to remove the function declaration
when removing the dlopen part of the code.

Signed-off-by: Edgar Gabriel <Edgar.Gabriel@amd.com>
(cherry picked from commit b4c55c443838da617c7fc9a2c86dc7b12c041a70)