timing framework: fix use of timers
Related to PR #11280

and issue #11213.

Turns out owing to the libopal-core.la refactor that we have to add in a bit
of plumbing to help the timing framework know when it can use native counters
vs gettimeofday.

With this patch, the timing framework now produces sensible time resuls
even prior to initialization of opal, i.e. when there are timers present in the
opal "util" initialization procedure.

Note that currently the timing points in ompi_mpi_init in main are messed up
so output isn't very useful.

A subsequent PR wherein the PMIx fence usage in the init procedure is cleaned
up will include refactored timing points.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>