MTL/OFI: Fix MPI_Cancel dereferences NULL pointer
This change postpones the use of ofi_req->comm until we are
sure it was initialized and we intend to use it.  Previously
attempting to cancel an IProbe or ISend would crash, even
though it should have been a no-op.

Additionally we must ensure ofi_req->type was properly
initialized, which it was not for a couple of SEND cases.

Signed-off-by: Luke Robison <lrbison@amazon.com>