MTL/OFI: Refactor SSend ACK generation and allow MRecv to process SSend's.
This commit refactors the generation of 0-byte ACKs from the body
of ompi_mtl_ofi_recv_callback so that it can be used by the
ompi_mtl_ofi_mrecv_callback function as well.

Previously the MRecv() callback did not generate the required ACKs,
and the sender would deadlock if MRecv'ing an SSend message.

Signed-off-by: Luke Robison <lrbison@amazon.com>