accelerator/cuda: Account for possibility that cuda will never be initialized
When the accelerator cuda component was first implemented,
if cuda was not initialized, the component would not be selected.
Therefore some logic was implemented with the assumption that
the initialization would have already happened.

After delayed initialization logic was added, the check
addr functions could be entered and would return an error if
cuda was not already initialized. Since cuda may never
be initialized, this patch adds some assumptions that the buffers
are host buffers if initialization does not pass.

Signed-off-by: William Zhang <wilzhang@amazon.com>