Request: fix race between registering callback and completing request
There is a subtle race condition where a thread registering
a callback in ompi_request_set_callback might see the request complete
and a thread completing the request might see the callback.
Both threads would then execute the callback. Similarly, there
may be situations where the callback is never executed because
no thread observes the two states at once. While the window
of opportunity is small, it seems to have occured in codes where
an application thread serves as progress thread.

This patch makes sure that
a) the callback is only ever executed only once by atomically setting
   the callback pointer if needed. This atomic is avoided when possible.
b) the callback is executed after all by cross-checking for the
   two states if a conflict may arise.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>