Fixing VERSION file for v5.0.0rc10
bot:notacherrypick

Open MPI v5.0.0 shared libraries are ABI compatible with v4.1.x with a
few subtle possible exceptions for Fortran.

In the rare case that you compile your application in such a way that
the size of an integer for C is different than the size of an integer
in Fortran, you'll need to rebuild and relink your application.

There are some additional Fortran API changes involving intents and
asyncs, along with changing some interfaces from named to unamed.

Resetting the age of all internal libraries to 0 for v5.0.0

Refer to https://docs.open-mpi.org/en/v5.0.x/version-numbering.html#shared-library-version-number for policy

Signed-off-by: Geoffrey Paulsen <gpaulsen@us.ibm.com>