coll/han: fix the count of allreduce_t3_task's lower bcast operation
The lower bcast operation of t3 task was applied to current segment.
Prior to this patch, the count of this operation is always seg_count,
which is wrong, because this operation can be applied to last segment
when there are mulitple segments. Therefore its count need to be
variable.

This patch addressed the issue by setting the count of this operaiton
to last_seg_count when it is applied on last segment.

Signed-off-by: Wei Zhang <wzam@amazon.com>