3rd-party/romio/mpl/configure.ac: fix AC 2.71 and m4 quoting issues
Directly copy and pasted
https://github.com/pmodels/mpich/blob/38552cfab7ac139754066e09f2a85e8b5a7d3df7/src/mpl/configure.ac#L451-L513
to replace the equivalent block here in OMPI's tree to fix some
Autoconf 2.71 portability issues and some m4 quoting issues.

Note: there are many other Autoconf 2.71 portability issues here in
the 3rd-party/romio341 tree that could also be fixed, but they only
generate warnings (vs. the blob that we back-ported, which fixes an
error that causes configure to abort).  Hence, we're not fixing them
-- we'll pick up those changes the next time we update ROMIO.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>