coll/han: call fallback functin when HAN module is disabled
This patch is to address:
    https://github.com/open-mpi/ompi/issues/11448

When Open MPI is compiled with CUDA support,
comm->c_coll->coll_xxx_module is coll_cuda_module and
HAN_LOAD_FALLBACK_COLLECTIVE is a no-op.

As a result, HAN's collective functions can be called
even if HAN has been disabled, which resulted an infinitely
recursive calling loop.

To address this issue, this patch make HAN's collective
fucntion to call fallback function when HAN module was
disabled.

Signed-off-by: Wei Zhang <wzam@amazon.com>