PMIx_Fences - remove two during MPI initialization
This patch removes redundant PMIx Fences in the initialization
procedure for MPI when using the World Process Model (WPM).
See chapter 11 sections 2 and 3 of the MPI-4 standard for a discussion
of the WPM and new Sessions model.

The patch does, however, require that what should have been a local operation
to support initialization of an MPI session, into a global one.
Note this does not disable the sessions feature but just restricts when it
will work at this point to use cases that are similar to MPI initialization
using the WPM.

Refactoring to make ompi_mpi_instance_init_common  purely local will require changes that would
be too impactive for the current state of the 5.0.0 release cycle.
See issue #11239.

Related to #11166

Signed-off-by: Howard Pritchard <howardp@lanl.gov>