Allow for optimized scale launches
Many environments do not face the problem of inconsistent
pml component selection. Yet the current pml implementation
requires that either a modex be executed so that each
process can check their pml selection against that of
rank=0, or that each process engage in a very expensive
all-to-all direct modex exchange of the pml selection.

Provide an opportunity for users and system admins to
bypass both of these methods by setting an MCA param
(via one of the several OMPI-supported methods) that
disables the check. This has significant launch
performance impact for most large-scale systems, but
we will leave the param set to default to executing
the check since those who need it are probably not
savvy enough to know that they do.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 767ba1a486d11ac73768a8aee24b825a99f9ae35)