Jenkins pr-builder.sh: Add "--bind-to none"
Similar to
https://github.com/open-mpi/ompi-scripts/commit/7dc912ef5c1542a5fe7e84d32f816c98a6439b78,
add "--bind-to none" to the mpirun command.

The CI tests on the AWS jenkins instance all run on instances with 1
core.  There's a special case in OMPI to bind to core for a 2 process
run for benchmarking reasons, and this was causing failures in mapping
because there aren't 2 cores on the instance.  So instead turn off
binding for these small tests.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>