btl/ofi: use libfabric 1.18 API if available
This patch try to use 1.18 API when it is available. This is because
1.18 API clearly define provider's CUDA API behavior to be that
provider can call CUDA API by default if application uses 1.18 API.

When using older version of API, some libfabric will not claim
support of FI_HMEM even if it is capable of supporting because
the provider does not know whether CUDA calls are permitted.

Signed-off-by: Wei Zhang <wzam@amazon.com>
(cherry picked from commit 9684337995cf9ca35855668381526a8630b1b2f8)