Pass NULL to the collective components for unused buffers.
We want the underlying algorithms to know when they can use a buffer as
a temporary, such as when one module delegate a collective to another.
Instead of allocating temporaries everywhere, we allow the upper level
module to handle the temporary buffers, and pass them into all the
others. Thus, we need to make sure that the user provided buffers are
screened such that they don't make it as temporaries.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>