configure.ac: check for MPI_T.3 (not MPI_T.5) for pre-built docs
We moved the MPI_T man page from section 5 to section 3 (where it
belongs), but forgot to update the configure check that uses that man
page as a sentinel to know if the pre-built docs are available.

Thanks to Orion Poplawski for noticing the problem.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>
(cherry picked from commit fa8885ed10fe0464309320cb1cebf57a3c847674)