pr-checks: add milestoner GitHub Action
The milestoner action adds a milestone to a pull request based on the base
branch name. If the base branch name matches "vNUMBER.NUMBER", look for all open
milestones matching this prefix. If any are found, apply the one with the
earliest due date. If none are found or if a milestone is already applied, do
nothing.

Signed-off-by: Joe Downs <joe@dwns.dev>
(cherry picked from commit 2040f91b7ffc492ab5133253208d20e00bd0d357)