catch exceptions thrown by sh step running pr-builder.sh
If the pr-builder fails, the sh step running it will throw an exception. This
exception prevents later steps from running. If this exception is raised, we
catch it so the job can properly finish. However, catching the exception will
set the overall job status to sucess, but we can manually set it back to a
failure.

Signed-off-by: Joe Downs <joe@dwns.dev>