add link to GitHub PR in build description on Jenkins
The "Details" link at the bottom of the GitHub PR page brings you to the
specific build run on Jenkins, which doesn't have a direct link back to the PR
on GitHub. We're using the description of the build to add a link back to the PR
on GitHub.

Signed-off-by: Joe Downs <joe@dwns.dev>
(cherry picked from commit ce98e00f2b)