Don't overwhelm the blockLength.
During the optimization process the datatype engine will merge
contiguous elements. When this happen adding two integers might result
in a "largher than int" value and the resulting datatype will be
incorrectly sized. Use size_t instead of int (this is correct because
the MPI standard mandates the block lengths to be non-negative numbers).

Thanks @albandil for finding the issue and providing a patch.

Fixes #11607.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>