coll/base: do not send and receive 0 byte message for linear alltoallv
Prior to this change, the linear algorithm of alltoallv will post
send and receive even when sendcount and recvcount is 0.

This patch make the algorithm to skip 0 byte send/receive.

This patch works because MPI standard has strict requirement on
collective message size.

According to MPI standard's Collectives Introduction and Overview
(6.1) section:

For collective operations, the amount of data sent must
exactly match the amount of data specified by
the receiver.

Signed-off-by: Wei Zhang <wzam@amazon.com>