mtl/ofi: Fix EP Naming Bug for fi_av_insert()
Endpoint names can never be longer than FI_NAME_MAX.  The size returned
from fi_getname() is specific to the endpoint it is being called on.

https://ofiwg.github.io/libfabric/v1.1.1/man/fi_cm.3.html

Libfabric expects the list of endpoint names to be contiguous in
memory, and deliminated by the NULL terminator. OMPI was allocating an
arrays of strings with size of its local endpoints name. OMPI was
putting the start of each endpoint name at the start of each string
in the array of strings. This leads to gaps/incomplete names if the
endpoint name that was being added is a different size than the endpoint
that had fi_getname() called on it. This bug will cause startup hangs b/c
libfabric will try to send to an endpoint which doesn't exist.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>