v5.0.x mtl/ofi: scope down ompi_mtl_ofi_wc to function level
Currently, ompi_mtl_ofi_wc is scoped to be thread level, but not function
level, so if calls to ompi_mtl_ofi_context_progress are nested,
the ompi_mtl_ofi_wc will be overwritten. In ompi_mtl_ofi_context_progress,
the callback of the ofi_req extracted from the cq entry can trigger progress
engine in the ssend path, in which the receiver will post a tsendmsg to sender as ack.
After the callback finishes, it falls back to the original progress engine
loop, then it could progress the cqe that is already finished in the callback.

Signed-off-by: Shi Jin <sjina@amazon.com>
(cherry picked from commit 9fcaac2564d58d386a10193011c9f971a5a4a138)