Use acquire/release barrier to synchronize completion of request and test/wait
Before completing the request we should make sure that all other pieces
are visible, the status. The release barrier should synchronize with
an acquire barrier before reading the status in test/wait.

This also removes a full memory barrier at the beginning of test functions
that does not seem to serve any purpose.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>