opal/ofi: fallback to process id as package rank when unbound
The previous change caused inconsistent package rank when the process is
unbound - two processes could return the same package rank.
If the process is unbound,  OPAL_MODEX_RECV_VALUE_OPTIONAL
throws an error, in which case and we should fallback to the process id
instead of attempting to return a local package rank.

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>