coll/tuned: handle non-float-associative ops in allreduce and reduce-scatter
Sum and prod are not float-associative, i.e., the rounding of floating
point numbers may lead to different results depending on the order of
operations. For allreduce and reduce-scatter, we should select algorithms
that guarantee the same result on all processes for these operations.

MPI Section 6.9.6:
> MPI requires that all processes from the same group participating
> in these operations receive identical results.

It appears that for allreduce only the `non-overlapping` variants guarantee
this, i.e., performing a reduce following by bcast/scatter.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>