Ask for the modex of all jobs connected.
The original code was merging the local modex with the modex of the
local processes on the first jobid. This lead to incorrect, and
mismatched, information among processes when joining multiple jobid
processes (such as on the second spawn merged).

This patch iterate over all the jobid on the list of "to connect"
processes and adds their information to the local modex.

Fixes #11724.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>