ofi: push part of commit 450c72d to v50x
part of commit 450c72d is needed on v50x.
Without it v50x branch can't be used on some systems with HPE SS11/CXI OFI provider.

this patch pulls in initialization of pmix_topology struct but not the sha
advancing on prrte and openpmix submodules.

related to #11841

bot:notacherrypick

Signed-off-by: Howard Pritchard <howardp@lanl.gov>