Fix inconsistent configure-time Fortran TKR check
This fixes the inconsistency between what is tested at the configure time (the
unnamed Fortran interface) and what is actually used in the source code (the
named Fortran interface).

As was mentioned in #9445, the unnamed interfaces are more forgiving when it
comes to TKR mismatches. Therefore, the current configure-time check is not
strict enough and is prone to giving false results (e.g. for NVHPC 23.3+). The
fix is to switch to the named Fortran interface in the configure script.

Note that the inconsistency was resolved in the main branch with #9445, which we
cannot cherry-pick because it potentially breaks the ABI compatibility.

Signed-off-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>