fortran: fix MPI_F_STATUS_SIZE
This commit is a partial cherry pick of 7fce2f30.  7fce2f30's original
commit message was:

    update MPI_F08_status type

    Make the C MPI_F08_status type definition match the updated
    mpi_f08 type(MPI_Status) definition.

    This fix the inconsistency introduced in open-mpi/ompi@98bc7af7d4dce3fe9106ffff39c0b129b9466d40

    Signed-off-by: Gilles Gouaillardet <gilles@rist.or.jp>

This commit only uses a portion of 7fce2f30 because we do not want to
also take the mpi_f08 Status part of that commit (because the mpi_f08
module does not exist on the v4.1.x branch).

Thanks to Github user @jprotze for identifying the issue and tracking
down the fix.

Signed-off-by: Jeff Squyres <jeff@squyres.com>
(cherry picked from commit 7fce2f3057d6baaba688c99ad5602407df21128d)